+++ Changelog.md
@@ -9,6 +9,7 @@
 - show random button only when graph-data is availible [#168](https://github.com/secure-77/Perlite/issues/168)
 - fixed some pop hover issues (use with 100%) and show content
 - implemented support for internal Markdown Links [#170](https://github.com/secure-77/Perlite/issues/170)
+- implemented hidden text feature [#160](https://github.com/secure-77/Perlite/issues/160)